refactor: Add react-select dependency and update create-post page
This commit adds the "react-select" dependency to the project and updates the create-post page to include a new feature of selecting tags and categories using the react-select component. The existing tags and categories are fetched from the server and displayed as options in the select dropdowns. This improves the user experience and makes it easier to select tags and categories when creating a new post.